requirements.txt
emoji==2.2.0
emoji-country-flag==1.3.1
geonamescache==1.5.0
pytest==7.2.0
pyfiglet==0.8.post1
python-dotenv==0.21.0